Posts/Feed for Games
Scheme and routes were setup for the posts.

A rudimentary system for viewing the posts successfully works. When going to a game page, the app retrieves all the posts related to that game. The posts are then displayed in the feed section of the game page. The user can scroll through the posts if the number of posts ends up exceeding the size alloted for the feed.