Profile dropdown and feature games carousel
There is now a dropdown for the profile on the right most side of the navbar. Currently this gives options for "Edit Profile" and "Settings" which are non-functional right now. There is also a logout option beneath these two which is funciontal. Featured games has also been updated so now it is a carousel that can be cycled through to view all the featured games.