Post details
When creating a post, the user now has the option of selecting which platform pertains to their post. The current options are: no platform, xbox, playstation, pc, steam. A field has been added to the game post schema to include information about what platform the post is about. Added icons for each of the platforms that are displayed at the top right of each post.