Reply count
Added a reply count attribute to the post schema that keeps track of how many replies there are on that post. The reply count is updated whenever a reply is added or deleted. The reply count is used on the game page and post page to display the amount of replies on that post.